Implemented the snode pool cache logic, PR comments
The Network object now manages an internal service node pool cache as well as the onion request paths needed for sending requests.

When calling `send_onion_request` it will check the snode pool (populating it if needed) and check for possible onion request paths (creating them if needed) before sending the request - these pool/path population approaches are blocking so the caller can trigger multiple requests and doesn't need to worry about multiple requests being triggered to populate the cache.

The Network object also has logic to persist and load this cache from disk (when initialised with a `cache_path`) in order to improve launch times after retrieving an initial cache.

Additionally the Network object exposes a couple of hooks so the client can register for updates to the onion request paths or the network connection status.