Remove unneeded specific values
We don't care about the specific numeric values underlying this enum, so
don't need to give them.  (It'll effectively be the same under the hood,
but by not giving them we signal that we just want different values, but
don't care what they are underneath, just that they are distinct).