Re-join disk writing thread during destruction
If we don't wait for the disk thread to exit then we could end up
destroying the Network instance, but the disk writing thread also
depends on that instance and so we could end up destroying its
referenced `this` object out from under it.

This tweaks it to store the std::thread in Network rather than detaching
it so that we can re-join it (which waits for it to finish) before we
complete the Network destruction.

This also renames the disk thread function slight to
`disk_write_thread_loop` (from `start_disk_write_thread`) because the
function doesn't actually *start* the thread: it is just the infinite
loop body that the thread runs forever.