Remove custom comparison for Address compare
Address (and RemoteAddress) already have a built-in `<` operator that
does pretty much what this custom comparator does.

(Although I found a weirdness in it that sorts the port in an unexpected
way -- https://github.com/oxen-io/oxen-libquic/pull/128 fixes that.
That fix isn't strictly necessary here, though, because here we just
need unique ordering, but the weird ordering in current libquic is still
unique).