Fixed an issue when merging the current config into itself
This change fixes an issue where when the user would merge a config message which is identical to the current config - there was existing logic to handle this case (no merge would occur and the existing config gets used) but it only applied to the `all_confs` vector and not the `all_hashes` vector. The result of this discrepancy would mean that the current hash could be added to the `_old_hashes` set, resulting in the current config messages getting deleted from the swarm.

This change avoids adding the hash to `_old_hashes` if it's the same as the current config hash.